Default exports + file rename for consistency
