+++ detox/src/utils/isRegExp.js
@@ -2,4 +2,6 @@ function isRegExp(obj) {
   return Object.prototype.toString.call(obj) === '[object RegExp]';
 }
 
-module.exports = isRegExp;
+module.exports = {
+  isRegExp,
+};
